requirements.txt
camelot_py==0.11.0
pandas==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
tqdm==4.67.1
